wifi: v1.2.12_20210906-r driver:29eae25d21eccad37858fd98af654954b06790f0 [1/1]
PD#OTT-25465

Problem:
wifi power low

Solution:
update type to fix low power issue

Verify:
verified ok

+++ project_w1/common/patch_fi_cmd.h
@@ -78,7 +78,7 @@ typedef struct tx_error_event
 typedef struct Cali_Param
 {
     unsigned char Cmd;
-    unsigned long version;
+    unsigned int version;
     unsigned short cali_config;
     unsigned char freq_offset;
     unsigned char htemp_freq_offset;
