+++ project_w1/vmac/wifi_hal_cmd.c
@@ -1723,11 +1723,11 @@ unsigned char get_s16_item(char *varbuf, int len, char *item, short *item_value)
     return 1;
 }
 
-unsigned char get_s32_item(char *varbuf, int len, char *item, long *item_value)
+unsigned char get_s32_item(char *varbuf, int len, char *item, int *item_value)
 {
     unsigned int n;
     char tmpbuf[60];
-    long *p = item_value;
+    int *p = item_value;
     int ret = 0;
     unsigned int pos = 0;
     unsigned int index = 0;
@@ -1755,7 +1755,7 @@ unsigned char get_s32_item(char *varbuf, int len, char *item, long *item_value)
                 while ((varbuf[pos] != 0) && (varbuf[pos] != ',') && (pos < len))
                     tmpbuf[n++] = varbuf[pos++];
 
-                *p++ = (long)simple_strtol(tmpbuf, NULL, 0);
+                *p++ = (int)simple_strtol(tmpbuf, NULL, 0);
             }
             while (varbuf[pos++] == ',');
 
@@ -1825,7 +1825,7 @@ unsigned char parse_cali_param(char *varbuf, int len, struct Cali_Param *cali_pa
 {
     unsigned short platform_verid = 0; // default: 0
     unsigned short cali_config[1] = {0};
-    unsigned long version[1] = {0};
+    unsigned int version[1] = {0};
     get_s32_item(varbuf, len, "version", version);
     cali_param->version = version[0];
     get_s16_item(varbuf, len, "cali_config", cali_config);
@@ -1859,7 +1859,7 @@ unsigned char parse_cali_param(char *varbuf, int len, struct Cali_Param *cali_pa
 #endif
     cali_param->platform_versionid = platform_verid;
 
-    printk("======>>>>>> txt version = %ld\n", cali_param->version);
+    printk("======>>>>>> txt version = %d\n", cali_param->version);
     printk("======>>>>>> cali_config = %d\n", cali_param->cali_config);
     printk("======>>>>>> freq_offset = %d\n", cali_param->freq_offset);
     printk("======>>>>>> htemp_freq_offset = %d\n", cali_param->htemp_freq_offset);
@@ -2213,7 +2213,7 @@ unsigned int hal_cfg_cali_param(void)
     wf5g_txpwr_param.Cmd = WF5G_TXPWR_PARAM_CMD;
     err = get_cali_param(&cali_param, &wf2g_txpwr_param, &wf5g_txpwr_param);
 
-    printk("calibration parameter: version %ld, config %d, freq_offset %d, tssi_2g %d, tssi_5g %d %d %d %d tx_en %d\n",
+    printk("calibration parameter: version %d, config %d, freq_offset %d, tssi_2g %d, tssi_5g %d %d %d %d tx_en %d\n",
             cali_param.version, cali_param.cali_config, cali_param.freq_offset, cali_param.tssi_2g_offset,
             cali_param.tssi_5g_offset[0], cali_param.tssi_5g_offset[1], cali_param.tssi_5g_offset[2], cali_param.tssi_5g_offset[3], cali_param.wftx_pwrtbl_en);
